# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduBridge API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface SuccessResponse {
  message: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserResponse {
  id: string;
  name: string;
  email: string;
  role: string;
}

export interface AuthResponse {
  token: string;
  user: UserResponse;
}

export interface Subject {
  id: string;
  title: string;
  description: string;
}

export interface CreateSubjectRequest {
  title: string;
  description: string;
}

export interface Content {
  id: string;
  subjectId: string;
  title: string;
  notes: string;
  videoUrl?: string;
}

export interface CreateContentRequest {
  subjectId: string;
  title: string;
  notes: string;
  videoUrl?: string;
}

export interface Question {
  id: string;
  question: string;
  options: string[];
  correctAnswer: number;
}

export interface Quiz {
  id: string;
  subjectId: string;
  questions: Question[];
}

export interface CreateQuizRequest {
  subjectId: string;
  questions: Question[];
}

export interface QuizSubmitRequest {
  subjectId: string;
  answers: number[];
}

export interface Result {
  id: string;
  userId: string;
  subjectId: string;
  score: number;
  total: number;
  percentage: number;
}

export interface ResultWithSubject {
  id: string;
  userId: string;
  subjectId: string;
  subjectTitle: string;
  score: number;
  total: number;
  percentage: number;
  createdAt: string;
}

export interface MarkCompleteRequest {
  contentId: string;
}

export interface ProgressResponse {
  completedLessons: string[];
}