# lib/api-zod/src/generated/types/progressResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduBridge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressResponse {
  completedLessons: string[];
}